Rename query files to *.ts
