+++ components/query/query-base.service.ts
