+++ components/query/query-vector.service.ts
